# testdata/openapi/valid/jolli.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "jolli",
    "version": "1.0.0",
    "description": "Automatically generated from code analysis"
  },
  "paths": {
...
    "/generate-from-repos": {
      "post": {
        "summary": "Create generate-from-repos",
        "description": "Extracted from backend/src/router/DocsiteRouter.ts",
        "tags": [
          "API"
        ],
        "responses": {
          "201": {
            "description": "Resource created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "repositories": {
                    "type": "array",
                    "items": {}
                  },
                  "name": {
                    "type": "string"
                  },
                  "displayName": {
                    "type": "boolean"
                  },
                  "visibility": {
                    "type": "boolean"
                  }
                },
                "required": []
              }
            }
          }
        }
      }
    },
    "/generate": {
      "post": {
        "summary": "Create generate",
        "description": "Extracted from backend/src/router/DocsiteRouter.ts",
        "tags": [
          "API"
        ],
        "responses": {
          "201": {
            "description": "Resource created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "integrationIds": {
                    "type": "number"
                  },
                  "name": {
                    "type": "string"
                  },
                  "displayName": {
                    "type": "boolean"
                  },
                  "visibility": {
                    "type": "boolean"
                  }
                },
                "required": []
              }
            }
          }
        }
      }
    },
...
  }
}